// src/domain/enums.ts
export enum PapelUsuario {
  ADMINISTRADOR = 'ADMINISTRADOR',
  OPERADOR_CADASTRO = 'OPERADOR_CADASTRO',
  GESTOR_ALMOXARIFADO = 'GESTOR_ALMOXARIFADO',
  AUDITOR = 'AUDITOR'
}

export enum StatusLote {
  RECEBIDO = 'RECEBIDO',
  EM_TRIAGEM = 'EM_TRIAGEM',
  TRIAGEM_CONCLUIDA = 'TRIAGEM_CONCLUIDA',
  ENCAMINHADO = 'ENCAMINHADO',
  FINALIZADO = 'FINALIZADO'
}

export enum TipoEquipamento {
  COMPUTADOR_MESA = 'COMPUTADOR_MESA',
  NOTEBOOK = 'NOTEBOOK',
  MONITOR = 'MONITOR',
  IMPRESSORA = 'IMPRESSORA',
  SERVIDOR = 'SERVIDOR',
  ROTEADOR = 'ROTEADOR',
  CABO_ESTRUTURADO = 'CABO_ESTRUTURADO',
  FONTE_ALIMENTACAO = 'FONTE_ALIMENTACAO'
}

export enum EstadoFisico {
  NOVO = 'NOVO',
  BOM_ESTADO = 'BOM_ESTADO',
  USADO_LEVE = 'USADO_LEVE',
  USADO_MODERADO = 'USADO_MODERADO',
  DANIFICADO_LEVE = 'DANIFICADO_LEVE',
  DANIFICADO_GRAVE = 'DANIFICADO_GRAVE',
  INSERVIVEL = 'INSERVIVEL'
}

export enum StatusRastreamento {
  AGUARDANDO_TRIAGEM = 'AGUARDANDO_TRIAGEM',
  EM_TRIAGEM = 'EM_TRIAGEM',
  AGUARDANDO_DESMONTE = 'AGUARDANDO_DESMONTE',
  EM_DESMONTE = 'EM_DESMONTE',
  PECAS_REAPROVEITADAS = 'PECAS_REAPROVEITADAS',
  MATERIAL_RECICLAVEL = 'MATERIAL_RECICLAVEL',
  DESCARTE_SEGURO = 'DESCARTE_SEGURO',
  BAIXA_DEFINITIVA = 'BAIXA_DEFINITIVA'
}